requirements.txt
altair==4.2.0
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.16.0
jupyter_client==7.4.8
jupyter_core==5.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nbclassic==0.4.8
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
packaging==22.0
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
plotly==5.10.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.4
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.14.0
Pympler==1.0.1
pymssql-linux
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.42
stack-data==0.6.2
streamlit==1.13.0
streamlit-nested-layout==0.1.1
tenacity==8.1.0
tensorflow==2.10.1
tensorflow-text==2.10.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
wincertstore==0.2
zipp==3.8.1
